requirements.txt
alembic==1.0.10
amqp==2.4.2
aniso8601==6.0.0
asn1crypto==0.24.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
kombu==4.5.0
Mako==1.0.10
MarkupSafe==1.1.1
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
redis==3.2.1
requests==2.22.0
retrying==1.3.3
shellescape==3.4.1
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.4
